navButtonを共通パーツに
