+++ lib/clock/clock_screen.dart
@@ -5,7 +5,8 @@ import 'package:buss_timetable/timetable/timetable_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../BussThemeData.dart';
+import '../buss_theme_data.dart';
+import '../widget/navigation_button.dart';
 
 class ClockRoute extends StatelessWidget {
   const ClockRoute({super.key});
@@ -38,7 +39,12 @@ class _ClockScreen extends ConsumerWidget {
                 child: TimelineSection(),
               ),
               Expanded(child: SizedBox()),
-              Center(child: _NavigateButton()),
+              Center(
+                child: NavigateButton(
+                  route: TimetableRoute(),
+                  child: Text('時刻表'),
+                ),
+              ),
               SizedBox(height: 16),
             ],
           ),
@@ -94,32 +100,6 @@ class ClockCard extends ConsumerWidget {
             size: 50,
           )
         ],
-      ),
-    );
-  }
-}
-
-class _NavigateButton extends StatelessWidget {
-  const _NavigateButton();
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 32),
-      child: SizedBox(
-        width: double.infinity,
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => const TimetableRoute()),
-            );
-          },
-          child: const Padding(
-            padding: EdgeInsets.symmetric(vertical: 16),
-            child: Text('時刻表'),
-          ),
-        ),
       ),
     );
   }
