+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:buss_timetable/BussThemeData.dart';
+import 'package:buss_timetable/buss_theme_data.dart';
 import 'package:buss_timetable/clock/clock_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
